# doc/api.yaml
openapi: 3.0.0
paths:
...
  /artist/{id}:
    get:
      description: Get single artist by id
      operationId: ArtistController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Returns if artist record exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Returns if artist ID is not valid UUID
        '404':
          description: Returns if artist record does't exist
      summary: Get single artist by id
      tags:
        - Artist
    put:
      description: Update artist information by UUID
      operationId: ArtistController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateArtistDto'
      responses:
        '200':
          description: Returns if artist has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '400':
          description: Returns if artist ID is not valid UUID
        '404':
          description: Returns if artist does not exist
      summary: Update artist information
      tags:
        - Artist
    delete:
      description: Delete artist from library
      operationId: ArtistController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if artist has been removed
        '400':
          description: Returns if artist ID is not valid UUID
        '404':
          description: Returns if artist does not exist
      summary: Delete artist
      tags:
        - Artist
  /track:
    get:
      description: Gets all library tracks list
      operationId: TrackController_findAll
      parameters: []
      responses:
        '200':
          description: Returns all tracks records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Track'
      summary: Get tracks list
      tags:
        - Track
    post:
      description: Add new track information
      operationId: TrackController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackDto'
      responses:
        '201':
          description: Returns if track has been created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Returns if request body does not contain required fields
      summary: Add new track
      tags:
        - Track
  /track/{id}:
    get:
      description: Get single track by id
      operationId: TrackController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Track ID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Returns if track record exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Returns if track ID is not valid UUID
        '404':
          description: Returns if track record does't exist
      summary: Get single track by id
      tags:
        - Track
    put:
      description: Update library track information by UUID
      operationId: TrackController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Track ID
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTrackDto'
      responses:
        '200':
          description: Returns if track has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
        '400':
          description: Returns if track ID is not valid UUID
        '404':
          description: Returns if track does not exist
      summary: Update track information
      tags:
        - Track
    delete:
      description: Delete track from library
      operationId: TrackController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Track ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if track has been removed
        '400':
          description: Returns if track ID is not valid UUID
        '404':
          description: Returns if track does not exist
      summary: Delete track
      tags:
        - Track
...
  /album/{id}:
    get:
      description: Get single album by id
      operationId: AlbumController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Returns if album record exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Returns if album ID is not valid UUID
        '404':
          description: Returns if album record does't exist
      summary: Get single album by id
      tags:
        - Album
    put:
      description: Update library album information by UUID
      operationId: AlbumController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAlbumDto'
      responses:
        '200':
          description: Returns if album has been updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '400':
          description: Returns if album ID is not valid UUID
        '404':
          description: Returns if album does not exist
      summary: Update album information
      tags:
        - Album
    delete:
      description: Delete album from library
      operationId: AlbumController_remove
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if album has been removed
        '400':
          description: Returns if album ID is not valid UUID
        '404':
          description: Returns if album does not exist
      summary: Delete album
      tags:
        - Album
...
  /favs/track/{id}:
    post:
      description: Add track to the favorites
      operationId: FavsController_addTrack
      parameters:
        - name: id
          required: true
          in: path
          description: Track ID
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: Returns if track has been added to the favorites
        '400':
          description: Returns if track ID is not valid UUID
        '422':
          description: Returns if track with ID does not exist
      summary: Add track to the favorites
      tags:
        - Favs
    delete:
      description: Delete track from favorites
      operationId: FavsController_removeTrack
      parameters:
        - name: id
          required: true
          in: path
          description: Track ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if track has been removed
        '400':
          description: Returns if track ID is not valid UUID
        '404':
          description: Returns if track is not favorite
      summary: Delete track from favorites
      tags:
        - Favs
  /favs/album/{id}:
    post:
      description: Add album to the favorites
      operationId: FavsController_addAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: Returns if album has been added to the favorites
        '400':
          description: Returns if album ID is not valid UUID
        '422':
          description: Returns if album with ID does not exist
      summary: Add album to the favorites
      tags:
        - Favs
    delete:
      description: Delete album from favorites
      operationId: FavsController_removeAlbum
      parameters:
        - name: id
          required: true
          in: path
          description: Album ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if album has been removed
        '400':
          description: Returns if album ID is not valid UUID
        '404':
          description: Returns if album is not favorite
      summary: Delete album from favorites
      tags:
        - Favs
  /favs/artist/{id}:
    post:
      description: Add artist to the favorites
      operationId: FavsController_addArtist
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      responses:
        '201':
          description: Returns if artist has been added to the favorites
        '400':
          description: Returns if artist ID is not valid UUID
        '422':
          description: Returns if artist with ID does not exist
      summary: Add artist to the favorites
      tags:
        - Favs
    delete:
      description: Delete artist from favorites
      operationId: FavsController_removeArtist
      parameters:
        - name: id
          required: true
          in: path
          description: Artist ID
          schema:
            format: uuid
            type: string
      responses:
        '204':
          description: Returns if artist has been removed
        '400':
          description: Returns if artist ID is not valid UUID
        '404':
          description: Returns if artist is not favorite
      summary: Delete artist from favorites
      tags:
        - Favs
info:
  title: Home Library Service
  description: Home music library service
  version: '1.0'
  contact: {}
...
servers: []
components:
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Freddie Mercury
        grammy:
          type: boolean
          example: false
      required:
        - id
        - name
        - grammy
...
    UpdateArtistDto:
      type: object
      properties:
        name:
          type: string
        grammy:
          type: boolean
    Track:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: The Show Must Go On
        artistId:
          type: string
          description: Refers to Artist
          nullable: true
          format: uuid
        albumId:
          type: string
          description: Refers to Album
          nullable: true
          format: uuid
        duration:
          type: integer
          description: In seconds
          example: 262
      required:
        - id
        - name
        - artistId
        - albumId
        - duration
    CreateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: Bohemian Rhapsody
        artistId:
          type: string
          description: Refers to Artist
          nullable: true
          format: uuid
        albumId:
          type: string
          description: Refers to Album
          nullable: true
          format: uuid
        duration:
          type: integer
          description: In seconds
          example: 355
      required:
        - name
        - artistId
        - albumId
        - duration
    UpdateTrackDto:
      type: object
      properties:
        name:
          type: string
          example: Bohemian Rhapsody
        artistId:
          type: string
          description: Refers to Artist
          nullable: true
          format: uuid
        albumId:
          type: string
          description: Refers to Album
          nullable: true
          format: uuid
        duration:
          type: integer
          description: In seconds
          example: 355
    Album:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Innuendo
        year:
          type: integer
          example: 1991
        artistId:
          type: string
          description: Refers to Artist
          nullable: true
          format: uuid
      required:
        - id
        - name
        - year
        - artistId
...
    UpdateAlbumDto:
      type: object
      properties:
        name:
          type: string
        year:
          type: integer
        artistId:
          type: string
          description: Refers to Artist
          nullable: true
          format: uuid
...